execution/requirements.txt
requests
python-dotenv
openai
supabase
psycopg[binary]
